Added gradle to the project
