--- duke.txt
@@ -1,2 +1,0 @@
-[ D ][ N ] eat by: 2018-01-01 
-[ E ][ N ] eat from: 2019-01-01 to: 2020-01-01
