+++ src/main/java/seedu/duke/Deadline.java
@@ -1,4 +1,4 @@
-package duke;
+package seedu.duke;
 
 import java.time.LocalDate;
