+++ src/main/java/seedu/duke/DukeException.java
@@ -1,4 +1,4 @@
-package duke;
+package seedu.duke;
 
 public class DukeException extends Exception{
     public DukeException(){
