+++ src/main/java/seedu/duke/Parser.java
@@ -1,4 +1,4 @@
-package duke;
+package seedu.duke;
 
 public class Parser {
