+++ src/main/java/seedu/duke/Task.java
@@ -1,4 +1,4 @@
-package duke;
+package seedu.duke;
 
 public class Task {
     protected String description;
